move PathExt into util module
